requirements.txt
controlnet_aux==0.0.7
cupy
cupy_cuda12x
decorator==4.0.2
diffusers
einops
Flask
Flask_Cors
gradio
huggingface_hub
imageio
lightning
moviepy
numpy
onnx==1.16.2
onnxruntime==1.19.2
opencv_python==4.10.0.82
pandas
peft
Pillow
Requests
safetensors
setuptools
torch>=2.0.0
transformers
typing_extensions
xformers
imageio
imageio[ffmpeg]
streamlit
streamlit-drawable-canvas